lowercase camera module
